+++ NEWS.md
@@ -1,3 +1,14 @@
+# weda 0.0.1
+
+* First main iteration of the app
+* Vignette and tools to help upload camera trap data  
+* Tools to help connect to database  
+* Shiny pp to view, download and upload camera trap data to the database  
+* Code for database views  
+* Functions to standardize names to VBA syntax  
+* Data dictionaries and schema for the camera trap database tables data('data_dictionary')
+
+
 # weda 0.0.0.9000
 
-* Created package with following functionality:
+* Created dev package